fix(stubs): correct optional parameters in function signatures
This commit corrects the signatures of several built-in functions in the stubs to accurately reflect that certain parameters are optional.

Previously, parameters like `$count` in `preg_replace` functions and `$length` in various `substr_*` functions were marked as required. This did not match the actual behavior of these functions in PHP, where they can be omitted.

The following functions have been updated to provide a default `null` value for their respective optional parameters:
-   `preg_replace`
-   `preg_replace_callback`
-   `preg_replace_callback_array`
-   `preg_filter`
-   `substr_count`
-   `strspn`
-   `strcspn`
-   `str_word_count`
-   `substr_compare`
-   `similar_text`

This change improves the accuracy of the stubs, preventing incorrect "Too few arguments" errors when these optional parameters are not provided.

Signed-off-by: azjezz <azjezz@protonmail.com>