fix(analyzer): correctly handle conditional exits in null coalesce expressions
This commit fixes a bug in the control flow analysis where an exiting expression (like `throw` or `exit`) on the right-hand side of a null coalesce operator (`??`) was treated as unconditional.

Previously, code like `$foo = $bar ?? throw new Exception()` would cause the analyzer to incorrectly mark all subsequent code as unreachable.

The analysis of the right-hand side of a `??` is now correctly handled as a conditional code path. An exit within the RHS no longer affects the control flow of the code following the expression, thus eliminating a class of "unreachable code" false positives.

Closes #448.

Signed-off-by: azjezz <azjezz@protonmail.com>